fix failed tests in yingchun gived list
