+++ python/paddle/fluid/tests/unittests/test_program_code.py
@@ -12,71 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-import time
-import unittest
-import sys
-from multiprocessing import Process
-import signal
+from __future__ import print_function
 
-import numpy
+import unittest
 
 import paddle.fluid as fluid
 import paddle.fluid.layers as layers
-from paddle.fluid.layers.io import ListenAndServ
-from paddle.fluid.layers.io import Recv
-from paddle.fluid.layers.io import Send
-import paddle.fluid.layers.ops as ops
-
-
-class TestProgram2Code(unittest.TestCase):
-    @unittest.skipIf(sys.platform == "win32",
-                     "Windows does not support distribution")
-    def test_print(self):
-        place = fluid.CPUPlace()
-        self.init_serv(place)
-        self.init_client(place, 9123)
-
-    def init_serv(self, place):
-        main = fluid.Program()
-
-        with fluid.program_guard(main):
-            serv = ListenAndServ("127.0.0.1:0", ["X"], optimizer_mode=False)
-            with serv.do():
-                out_var = main.global_block().create_var(
-                    name="scale_0.tmp_0",
-                    psersistable=True,
-                    dtype="float32",
-                    shape=[32, 32])
-                x = layers.data(
-                    shape=[32, 32],
-                    dtype='float32',
-                    name="X",
-                    append_batch_size=False)
-                fluid.initializer.Constant(value=1.0)(x, main.global_block())
-                ops._scale(x=x, scale=10.0, out=out_var)
-
-        print(main)
-
-    def init_client(self, place, port):
-        main = fluid.Program()
-        with fluid.program_guard(main):
-            x = layers.data(
-                shape=[32, 32],
-                dtype='float32',
-                name='X',
-                append_batch_size=False)
-            fluid.initializer.Constant(value=2.3)(x, main.global_block())
-            get_var = main.global_block().create_var(
-                name="scale_0.tmp_0",  # server side var
-                dtype="float32",
-                persistable=False,
-                shape=[32, 32])
-            fluid.initializer.Constant(value=2.3)(get_var, main.global_block())
-            Send("127.0.0.1:%d" % port, [x])
-            o = Recv("127.0.0.1:%d" % port, [get_var])
-
-        print(main)
 
 
 class TestProgramToReadableCode(unittest.TestCase):
