+++ python/paddle/fluid/tests/unittests/test_translated_layer.py
@@ -94,7 +94,7 @@ class TestTranslatedLayer(unittest.TestCase):
             batch_size=BATCH_SIZE,
             shuffle=True,
             drop_last=True,
-            num_workers=2)
+            num_workers=0)
 
         # train
         train(self.layer, self.loader, self.loss_fn, self.sgd)
